Revert "Revert "Review/cv 1214 uploads (#317)" (#318)"
This reverts commit 41311c4751968533003da05c96e0fb9265f410d8.
